refactor: move VimarStatusSensor to binary_sensor module
VimarStatusSensor extends BinarySensorEntity and is only instantiated
when platform == DEVICE_TYPE_BINARY_SENSOR, so it belongs alongside the
other binary sensor classes — not in the platform-agnostic base-entity
module.

Side effects:
- vimar_setup_entry() drops its DEVICE_TYPE_BINARY_SENSOR special case;
  binary_sensor.async_setup_entry() now instantiates the status sensor
  directly before delegating to vimar_setup_entry().
- vimar_entity.py no longer imports BinarySensorEntity,
  BinarySensorDeviceClass, CONF_VERIFY_SSL, or DEVICE_TYPE_BINARY_SENSOR.

No behavior change: the status sensor is still added exactly once per
config entry, with the same unique_id, device_info, and update() logic.
