+++ custom_components/vimar/binary_sensor.py
@@ -10,6 +10,7 @@ from homeassistant.components.binary_sensor import (
     BinarySensorEntity,
 )
 from homeassistant.config_entries import ConfigEntry
+from homeassistant.const import CONF_VERIFY_SSL
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
@@ -78,11 +79,15 @@ async def async_setup_entry(
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up the Vimar BinarySensor platform."""
+    coordinator: VimarDataUpdateCoordinator = hass.data[DOMAIN][entry.entry_id]
+
+    # Connection status sensor (always present, not in device tree)
+    async_add_entities([VimarStatusSensor(coordinator)], True)
+
     # Standard Vimar binary sensors (monostable switches)
     vimar_setup_entry(VimarBinarySensor, CURR_PLATFORM, hass, entry, async_add_entities)
 
     # SAI2 zone sensors (alarm physical sensors)
-    coordinator: VimarDataUpdateCoordinator = hass.data[DOMAIN][entry.entry_id]
     vimarproject = coordinator.vimarproject
 
     if vimarproject is None or vimarproject.sai2_zones is None:
@@ -105,6 +110,59 @@ async def async_setup_entry(
         existing.extend(zone_entities)
     else:
         coordinator.devices_for_platform[CURR_PLATFORM] = zone_entities
+
+
+class VimarStatusSensor(BinarySensorEntity):
+    """Representation of Vimar connection status sensor."""
+
+    _coordinator: VimarDataUpdateCoordinator
+    _attr_should_poll = True
+
+    def __init__(self, coordinator: VimarDataUpdateCoordinator):
+        """Initialize the sensor."""
+        self._coordinator = coordinator
+        vimarconfig = coordinator.vimarconfig
+        conn = coordinator.vimarconnection._connection
+        self._attr_name = "Vimar Connection to " + str(conn._host) + ":" + str(conn._port)
+        self._attr_device_class = BinarySensorDeviceClass.CONNECTIVITY
+        self._attr_extra_state_attributes = {
+            "Host": conn._host,
+            "Port": conn._port,
+            "Secure": conn._schema == "https",
+            "Verify SSL": conn._schema == "https" and vimarconfig.get(CONF_VERIFY_SSL),
+            "Vimar Url": f"{conn._schema}://{conn._host}:{conn._port}",
+            "Certificate": conn._certificate,
+            "Username": conn._username,
+            "SessionID": coordinator.vimarconnection._session_id,
+        }
+        self._attr_is_on = False
+
+    @property
+    def unique_id(self):
+        """Return the ID of this device."""
+        prefix = self._coordinator.entity_unique_id_prefix or ""
+        if len(prefix) > 0:
+            prefix += "_"
+        return DOMAIN + "_" + prefix + "status"
+
+    @property
+    def device_info(self) -> DeviceInfo:
+        """Return device information."""
+        return DeviceInfo(
+            identifiers={(DOMAIN, self._coordinator.entity_unique_id_prefix or "", "status")},  # type: ignore[arg-type]
+            name="Vimar WebServer",
+            model="Vimar WebServer",
+            manufacturer="Vimar",
+        )
+
+    def update(self):
+        """Fetch new state data for the sensor.
+
+        FIX #10: this method is called by HA on the executor thread
+        (because _attr_should_poll = True and update() is synchronous),
+        so the blocking is_logged() network call is safe here.
+        """
+        self._attr_is_on = self._coordinator.vimarconnection.is_logged()
 
 
 class VimarBinarySensor(VimarEntity, BinarySensorEntity):
